helpers: remove the windows non-ascii workaround again
Removing the tempfile workaround for non-ascii filenames on Windows as
it is terribly inefficient in a multiprocessing environment.
Each separate process initialises its own PdfContext. If every single
process creates a temporary copy of the input file, this extremely harms
performance, in particular for large files.
I think we need to rely on users to check their file name and if
necessary create a temporary file before starting a multiprocessing pool
so the tempfile is created only once.