readme: shorten non-ascii section due to doubts about correctness
What I wrote about concurrent access might have been wrong.

I think it should be possible to access the same file concurrently
without the file name, by reading the data into a bytes object. The
initial mistake was to acquire the file descriptor in each process,
which does not work. Instead, the data could be read once and then be
passed to each process.

Whether we have access to the file descriptor in each job or not
shouldn't make any difference in terms of performance as we have to load
the full data into memory anyway in order to forward it to PDFium.

The problem is that loading the whole data into memory is not
practicable for (very) large files, so the tempfile workaround might be
preferable anyway. The best, though probably most complicated to
implement solution might be to check file size and then decide whether
to use a tempfile or load the data into memory.

I'll have to think about this more thoroughly and maybe add a test.