move version extraction away from `attr:`
I'm not sure about the cause since it worked previously, but creating
foreign wheels was broken for me because an import was attempted by
`attr:` to get the version. In the course of the imports, the generated
bindings file attempted to load the binary, so setup broke with an
exception.
This is now fixed by reading the version file literally without
importing anything. It had already been on my todo-list anyway.