nativeopener: add precautions to protect the reader callback object
I'm not sure how different versions or implementations of Python treat
unused variables. The LoaderData attribute `reader_instance` is not
accessed on the Python side, so a heavily optimising interpreter might
wrongly identify it as an unused object to remove, which would be very
disadvantageous for us.
I can't tell whether this is a realistic situation, but let's better be
safe than sorry.
This change adds a no-op call that accesses the object on
`LoaderData.close()`, to enforce that it remains available until then.