Move `PdfContext` back into opener
`PdfContext` should not be defined in `classes` as this induces a
potential circularity in the importer logic. The backend functions for
`PdfDocument` already need `PdfContext`, so we may not define both in
the same file.
The code worked anyway and the test suite passed, probably because the
Python interpreter has some intelligent heuristics to automatically
resolve said circularity. Nevertheless, `PdfContext` should really be
moved into `opener` to make the arrangement more logical and prevent
conflicts.