setup_base: use regular names for temporary tags
Each package contains a text file `WHEEL` in dist-info that stores the
tag it was generated with, so we should probably be using something
meaningful, instead of just a temporary name.
Fortunately, it looks like `WHEEL` is purely descriptive, and what is
relevant for the decision of pip which package to download remains the
file name.