readme: remove `FPDF_LoadMemDocument()` section
I now have a good guess why this weird behaviour might be happening.
The PDFium docs mandate that "the memory buffer must remain valid until
the document is closed". This means the cause for all the trouble could
be Python automatically starting to delete the byte string after it has
been passed to `FPDF_LoadMemDocument()`, for the handle variable to the
data is not accessed later. Presumably we could use the function in the
same way as `FPDF_LoadCustomDocument()` by setting up a data holder
object to keep the variable until the PDF is closed.