Considerably simplify the build script
Removing the nativebuild option as it's too awkward and probably
incomplete. It was difficult to test, too.
Users on platforms not supported by the default build infrastructure
should consult the PDFium mailing list. That said, it might be possible
to deploy a custom clang toolchain with `clang_base_path`, or even use
an entirely different build system, as LibreOffice do.
I hope that mainstream Linux distributions finally create an own package
for PDFium. This would save disk space, and could help avoid a lot of
development trouble.