Improve opening
The previous way of opening had issues with rendering integration:
If a byte buffer is given as input, it can't be used concurrently. We
can, however, acquire multiple byte buffers for the same file.

As we want to be able to open files in any way we like, the different
access strategies need to be implemented internally in `PdfDocument`.

This patch implements said functionality using a `FileAccess` enum and
three different input slots: `self._orig_input`, `self._actual_input`
and `self._rendering_input`.
`self._orig_input` is the original input provided by the caller.
`self._actual_input` at first refers to the same object as
`self._orig_input`. As the different file access strategies are taken
into account for string input, `self._actual_input` may be substituted
accordingly.
`self._rendering_input` are the sources that will be used for rendering,
and they are `None` at first. When the rendering method is called, it
selects its input source based on `self._orig_input`. Document handles
are saved implicitly, byte buffers are read into memory. Other input is
taken as-is. If `self._rendering_input` is set already, it will be
re-used and no input selection done.

The new features are not test-integrated yet!