Update planned changes
Use more vague wording for possible rendering parameter changes.
After thinking about this for a while, I feel rather opposed to the idea
of replacing all the mentioned parameters with a generic flag option.
Providing separate parameters makes it easier to define defaults and is
a bit more pythonic.
However, the current implementation for `no_antialias` isn't nice at
all, so I'd like to remove that parameter.
Nevertheless, adding an option for additional flags appears to be a good
idea, so that all PDFium options are available to the caller.