build: use minimal checkout configuration
This recently added option helps speed up the build.
See
https://pdfium.googlesource.com/pdfium/+/a2ccb26b273ab4d94db117ce2e85263b9e4628c6

While this definitely is a performance improvement, the situation is
still not optimal. Solely checking out the repository with git and then
using a custom build system (e. g. cmake, meson) with system-provided
binaries and libraries would be a lot better.