Add tasks concerning uninitialized struct fields
CC #129

I'm wondering if Python/ctypes is automatically zeroing struct fields
for us, or if we still need to care about this. May need to look it up.
But better be safe than sorry in any case.

With respect to FPDF_FORMFILLINFO, it provides so many fields that I
don't feel like assigning values to each and every one is reasonable,
but let's see.