Allow use of pixel format BGRX/RGBX
This format can still be useful for callers, e. g. if needed by the
target library, or if transparency is desired when post-processing the
image.
For instance, `PIL.Image.frombuffer()` can only reference the given
buffer for certain formats, and needs to create a copy otherwise.
RGB is not among these, but RGBX is (cf. PIL.Image._MAPMODES).
However, if this would actually lead to better performance remains to be
checked.