Finally fix #164
There seems to be an issue specific to the "fork" method and buffer
input.

With concurrent.futures input, it randomly crashes while processing.
With multiprocessing, it works with warnings, but we get a deadlock when
all jobs have finished.

"spawn" or "forkserver" work with both multiprocessing and
concurrent.futures.

We currently use "spawn". "forkserver" may be better, but we're not sure
about its availability.