POC patch to handle mmap as ctypes array
I think handling mmap as ctypes array (FPDF_LoadMemDocument64) may be
better than as buffer (FPDF_LoadCustomDocument) to avoid unnecessary
callbacks.

The way I understand mmap is that we can effectively treat it like
regular memory without the file actually being fully in memory (i. e.
the memory is a virtual view of a buffered stream, managed by the OS -
access of the memory span maps to that stream).