Rework safe_extract() code in the lights of PEP 706
The advantage of the new implementation is that it uses more elegant
strategies as soon as available. The logical drawback is that it's very
Python version conditional, making it harder to test.

I think the PEP was created very shortly after I wrote the previous
code. Pity I wasn't notified earlier.