Tighten ctypes member filter
In the (theoretical) case that bindings define differently a variable
included from the ctypes namespace, it should not be excluded.

However, it is impossible to detect deliberate assigns of a ctypes
member name to the same value (e.g. primitives like True or False). This
is an irrelevant edge case, though.
```
.. in ctypes:
common_attribute = True

.. in pypdfium2:
from ctypes import *
common_attribute = True
.. -> impossible to distinguish, would need manual inclusion
```
(using `..` for comments instead of hashtag because of
https://github.com/git-cola/git-cola/issues/1330)