fix a comment
There were some inaccuracies here - if V_PYPDFIUM2 is not in the yaml
file, then we also can't get it with load_data_file().

It remains to be seen how much yaml can do and how we would integrate
that into the API.

If it works out, then we may also be able to annotate the data/ cache in
a cleaner way instead of separate files for each conditional.