Move raw bindings to separate module (CC #256)
Enable the caller to select the modules to install using the env var
`$PYPDFIUM_MODULES`. This is an important pre-requisite for conda
packaging.

That said, it is just cleaner to isolate the raw bindings in a separate
module, so callers may use that without any automatic init from helper.
It even allows not to install helpers at all if you don't use them.
Similarly, it would allow to install only the helpers and plug in a
pypdfium2_raw supplied otherwise.

However, note that version integration is now really polluted (1 version
file for two separate modules - threatens desync).
In the future, we want to split this in two separate version files and
probably also use a yaml delegate for r/w instead of working directly
with the python file.