Use default src include if all modules given
If doing an sdist with the explicit modules dict, this has the
implication of pypdfium2.egg-info moving from src/ to the project root.

Not sure if this is desired, so better stick with the previous state.
The individual module code passage is used only with wheels, not with
sdist, and wheels do not seem to have .egg-info.