Simplify conda instructions
I believe it may be enough to just add the -c's on install and not
necessary to actually add the channels.

However, I haven't checked this very thoroughly. Let's just hope an
update won't mess up thing and try to replace pdfium-binaries or
something.

If it doesn't work we can always add back the tighter instructions.