Add comment regarding to_full()
'#273' gave us the hint that chromium ls-remote can be fairly expensive.
It may be better to use the pdfium-binaries VERSION file where
available. We might want to pick this idea.

Note, this does not mean it would be straightforward to take an offline
archive. It'd still need a new target, new env vars, handling down
parameters, changing headers integration, etc.
Since this is a very niche use case, it's cleaner to let callers do the
preprocessing IMO (see #274).

Thanks-to: nh2

(PR in quotes to work around
https://github.com/git-cola/git-cola/issues/1330)