conda/recipes: do not wrap env vars in quotes after all
conda produces something like this
```
The environment variable 'PDFIUM_PLATFORM' is being passed through with
value '"prepared!system:6150"'
```
So seemingly the quotes are not stripped yet when this print is made.
The result was still correct, so probably they're still stripped
somewhere at some point, but it seems we might be safer off not adding
them in the first place after all.