add back srcinfo to bindings file
--no-srcinfo temporarily removed in ctypesgen fork to reduce complexity,
due to issues in newlines management.

I suppose the diff space should be affordable if we only do monthly
releases (e.g. compared to test files we have in the repo - these should
be moved to a submodule in the future).
Also, the info is rather useful so we can directly go to a header
without having to r-grep for the symbol first.