Presumably work around pdfium bug 2112
I am however reluctant to make a release with this as it would wreak
breakage on downstreams that use a non-default bitmap maker.
Which would be unusual, but legit API use.

Also the issue seems to be very circumstance-specific and the cause in
the control flow is not yet logically understood (although we seem to
have narrowed down the commit and relation to stride), so I am not
certain if this truly works around all cases.