Remove PdfPosNormalizer experiment
The fact that the object has to be re-created in response to any change
in page geometry renders it basically unusable IMO.

The whole approach was bad practice anyway. The correct way to do this
would be to patch pdfium with new APIs that don't do the raster rounding
and origin translation.

FWIW, I made a private copy of the code.