PdfPage.get_objects(): don't register objects as kids
This was especially problematic as weakrefs are not cleaned up when the
object in question is closed/collected, so we potentially store many
dead pointers.
Imagine a caller invoking get_objects() repeatedly for iterating and a
page handle living for a long time afterwards - that somewhat resembles
a memory leak.