readme: improve raw api
Avoid declaring UTF-16 as "2 bytes per character", because a visual
character could be composed of a surrogate pair of 4 bytes.
So this is not the number of visual characters, but the number of units,
where the number of bytes per unit corresponds to the size of the data
type used.