cli/render: simplify --linear default
The current conditional default doesn't make too much sense, as a
non-jpg format probably means png, and png probably means cv2, and
importing that on N processes is a major slowdown.

The simplest compromise that came to mind was to just use a static
number. In this case, I chose for 4.