Take a step back on libname guessing
These config vars seem to be quite build/platform-specific, so they are
not overly useful for the "unknown host" case. e.g. on Windows, the
sysconfig module doesn't seem to tell us the libname pattern at all (as
of python 3.9).

Perhaps some other project has a clue how to get the libname pattern in
a portable way...