build_native: build abseil as separate shared library
For now, always use a separate DLL for abseil for simplicity.

If we want to provide a --single-lib option in the future, it might be
preferable to use the same strategy as in build_toolchained (e.g. if
more dependencies pop up).

It's good to know we can also choose whether to bundle or not to bundle
on a per-library basis, though.