req/docs: bump sphinx requirement
'include-read' requires sphinx >= 7.2.5, as per
https://github.com/executablebooks/MyST-Parser/issues/845#issuecomment-3060806840

Take out two other requirements as they are probably outdated. When
installing sphinx and myst, pip should implicitly install the right
version of sphinx-rtd-theme.
Actually take out docutils completely as it's not a direct dependency of
ours, so again, it should be installed implicitly, at the right version.