docs/conf.py: fix nasty or-expression blunder
A Python or-expression yields the *value* of the first sub-expression
that evaluates to True.

This means, if we are on release and the last expression is reached,
have_changes would evaluate to next_changelog itself, rather than its
boolean value.
So we would end up embedding the actual changelog_staging content in the
RST prolog, meaning `have_changes` doesn't work correctly, and every doc
page is garbled with a mis-formatted changelog at first.

This commit fixes this blunder.