docs/conf.py: actually comment out rst_prolog entirely
We never use `|have_changes|` in content, so not needed anymore.
For the `.. ifconfig` directive, the `app.add_config_value(...)` bit is
enough.

We did use the predecessor variable in content though (I think it was
the branch name or something).

Anyway, it's still good to know this is possible, so only comment it
out.